+++ packages/expo-router/src/react-navigation/core/__tests__/BaseNavigationContainer.test.ios.tsx
@@ -244,7 +244,7 @@ test('handle dispatching with ref', () => {
   });
 });
 
-test('handle resetting state with ref', () => {
+test('handles resetting to a complete state with ref', () => {
   const TestNavigator = (props: any) => {
     const { state, descriptors, NavigationContent } = useNavigationBuilder(MockRouter, props);
 
@@ -287,12 +287,18 @@ test('handle resetting state with ref', () => {
   render(element).update(element);
 
   const state = {
+    stale: false as const,
+    type: 'test',
+    key: 'navigator-3',
     index: 1,
+    routeNames: ['foo', 'foo2', 'bar', 'baz'],
     routes: [
       {
         key: 'baz',
         name: 'baz',
         state: {
+          stale: false as const,
+          type: 'test',
           index: 0,
           key: '4',
           routeNames: ['qux2', 'lex2'],
@@ -312,30 +318,29 @@ test('handle resetting state with ref', () => {
 
   expect(onStateChange).toHaveBeenCalledTimes(1);
   expect(onStateChange).toHaveBeenLastCalledWith({
+    stale: false,
+    type: 'test',
     index: 1,
-    key: '0',
+    key: 'navigator-3',
     routeNames: ['foo', 'foo2', 'bar', 'baz'],
     routes: [
       {
         key: 'baz',
         name: 'baz',
-        params: undefined,
         state: {
-          index: 0,
-          key: '1',
-          routeNames: ['qux2', 'lex2'],
-          routes: [
-            { key: 'qux2', name: 'qux2', params: undefined },
-            { key: 'lex2', name: 'lex2', params: undefined },
-          ],
           stale: false,
           type: 'test',
+          index: 0,
+          key: '4',
+          routeNames: ['qux2', 'lex2'],
+          routes: [
+            { key: 'qux2', name: 'qux2' },
+            { key: 'lex2', name: 'lex2' },
+          ],
         },
       },
-      { key: 'bar', name: 'bar', params: undefined },
+      { key: 'bar', name: 'bar' },
     ],
-    stale: false,
-    type: 'test',
   });
 });
 
@@ -501,6 +506,7 @@ test('emits state events when the state changes', () => {
   expect(listener).toHaveBeenCalledTimes(1);
   expect(listener.mock.calls[0]![0].data.state).toEqual({
     stale: false,
+    type: 'test',
     index: 1,
     key: 'navigator-3',
     routeNames: ['foo', 'bar', 'baz'],
@@ -517,6 +523,7 @@ test('emits state events when the state changes', () => {
   expect(listener).toHaveBeenCalledTimes(2);
   expect(listener.mock.calls[1]![0].data.state).toEqual({
     stale: false,
+    type: 'test',
     index: 2,
     key: 'navigator-3',
     routeNames: ['foo', 'bar', 'baz'],
@@ -954,6 +961,7 @@ test('works with state change events in independent nested container', () => {
       { key: 'lex-0', name: 'lex', params: undefined },
     ],
     stale: false,
+    type: 'test',
   });
 
   expect(ref.current?.getRootState()).toEqual({
@@ -965,6 +973,7 @@ test('works with state change events in independent nested container', () => {
       { key: 'lex-0', name: 'lex', params: undefined },
     ],
     stale: false,
+    type: 'test',
   });
 });